# backend/app/api/moments.py
from typing import Annotated, Literal

from fastapi import APIRouter, Depends, Header, Query, Request, Response
from pydantic import BaseModel, ConfigDict, Field

from app.core.config import Settings
from app.core.errors import AppError
from app.modules.identity.tokens import TokenService
from app.modules.moments.service import MomentsService
from app.modules.moments.media import MomentMediaService


class Strict(BaseModel):
    model_config = ConfigDict(extra="forbid")


class CreateMoment(Strict):
    text: str = Field(default="", max_length=5000)
    visibility: Literal["PUBLIC", "FRIENDS", "INCLUDE", "EXCLUDE", "SELF"]
    image_urls: list[str] = Field(default_factory=list, max_length=9)
    include_user_ids: list[str] = Field(default_factory=list)
    exclude_user_ids: list[str] = Field(default_factory=list, max_length=30)
    include_tag_ids: list[str] = Field(default_factory=list, max_length=30)
    exclude_tag_ids: list[str] = Field(default_factory=list, max_length=30)
    location: str | None = None
    link_url: str | None = None


class DraftPayload(Strict):
    payload: dict = Field(default_factory=dict)

class Comment(Strict):
    text: str = Field(min_length=1, max_length=1000)
    parent_id: str | None = None


class Preferences(Strict):
    history_range: Literal["ALL", "SIX_MONTHS", "ONE_MONTH", "THREE_DAYS"]
    personalized_recommendations: bool
    cover_url: str | None = None


class Report(Strict):
    reason_code: str = Field(min_length=1, max_length=100)

class BeginUpload(Strict):
    file_name: str = Field(min_length=1, max_length=255)
    mime_type: str = Field(min_length=1, max_length=100)
    byte_size: int = Field(gt=0)


class SetCover(Strict):
    upload_id: str = Field(min_length=1, max_length=36)


def create_moments_router(settings: Settings, factory, *, avatar_storage=None):
    router = APIRouter(prefix="/moments", tags=["moments"])
    service = MomentsService(factory, avatar_storage=avatar_storage)
    media = MomentMediaService(factory, storage=avatar_storage)
    tokens = TokenService(factory, jwt_secret=settings.jwt_secret or "development-jwt-secret-at-least-thirty-two-bytes", jwt_issuer=settings.jwt_issuer, require_session_claims=settings.environment != "test")

    def actor(authorization: Annotated[str | None, Header()] = None):
        if not authorization or not authorization.startswith("Bearer "):
            raise AppError(code="AUTH_REQUIRED", message="需要登录", status_code=401)
        return str(tokens.decode_access_token(authorization[7:])["sub"])

    @router.post("", status_code=201)
    def create(body: CreateMoment, idempotency_key: Annotated[str, Header(alias="Idempotency-Key")], user=Depends(actor)):
        # 直接返回新建动态的 DTO：detail() 仅对 PUBLISHED 可见，会令
        # 带图动态（PENDING_REVIEW 待审状态）发布时误报"动态不存在"。
        row = service.create(user, body.model_dump(), idempotency_key)
        from app.modules.moments.models import Moment

        with factory() as session:
            moment = session.get(Moment, row.id)
            return service.dto(session, moment, user)

    @router.get("/feed")
    def feed(mode: Literal["recommended", "latest"] = "recommended", cursor: str | None = None, limit: Annotated[int, Query(ge=1, le=50)] = 20, user=Depends(actor)):
        return {**service.feed(user, mode=mode, cursor=cursor, limit=limit), "mode": mode}

    @router.get("/search")
    def search(q: Annotated[str, Query(min_length=1, max_length=100)], cursor: str | None = None, limit: Annotated[int, Query(ge=1, le=50)] = 20, user=Depends(actor)):
        return service.feed(user, q=q, cursor=cursor, limit=limit)

    @router.get('/draft')
    def get_draft(user=Depends(actor)):
        return service.draft(user)

    @router.put('/draft')
    def save_draft(body: DraftPayload, user=Depends(actor)):
        return service.save_draft(user, body.payload)

    @router.delete('/draft', status_code=204)
    def delete_draft(user=Depends(actor)):
        service.delete_draft(user); return Response(status_code=204)

    @router.get('/ads')
    def ads(user=Depends(actor)):
        return {'items': service.native_ads()}

    @router.get('/notifications')
    def notifications(user=Depends(actor)):
        return {'items': service.notifications(user)}

    @router.get('/notifications/unread-count')
    def notification_unread_count(user=Depends(actor)):
        return {'count': service.notification_unread_count(user)}

    @router.post('/notifications/read', status_code=204)
    def mark_notifications_read(ids: list[str], user=Depends(actor)):
        service.mark_notifications_read(user, ids)
        return Response(status_code=204)

    @router.get('/users/{user_id}')
    def personal_timeline(user_id: str, user=Depends(actor)):
        return {'items': service.personal_timeline(user, user_id)}

    @router.get("/preferences")
    def preferences(user=Depends(actor)):
        return service.preferences(user)

    @router.put("/preferences")
    def update_preferences(body: Preferences, user=Depends(actor)):
        return service.preferences(user, body.model_dump())

    @router.post("/media/uploads", status_code=201)
    def begin_upload(body: BeginUpload, idempotency_key: Annotated[str, Header(alias="Idempotency-Key")], user=Depends(actor)):
        row = media.begin(user, body.file_name, body.mime_type, body.byte_size, idempotency_key)
        return {"id": row.id, "upload_url": f"/api/v1/moments/media/uploads/{row.id}/content", "expires_at": row.expires_at}

    @router.post("/media/uploads/{upload_id}/complete")
    def complete_upload(upload_id: str, idempotency_key: Annotated[str, Header(alias="Idempotency-Key")], user=Depends(actor)):
        row = media.complete(user, upload_id)
        media_url = avatar_storage.signed_read_url(row.object_key, 300) if row.status == "COMPLETED" and avatar_storage else f"media://{row.object_key}"
        return {"id": row.id, "status": row.status, "media_url": media_url}

    @router.put("/media/uploads/{upload_id}/content", status_code=204)
    async def put_upload_content(upload_id: str, request: Request, content_type: Annotated[str | None, Header(alias="Content-Type")] = None, user=Depends(actor)):
        content = bytearray()
        async for chunk in request.stream():
            content.extend(chunk)
        media.put_content(user, upload_id, bytes(content), (content_type or "").partition(";")[0].strip().casefold())
        return Response(status_code=204)

    @router.post("/cover/uploads", status_code=201)
    def begin_cover_upload(body: BeginUpload, idempotency_key: Annotated[str, Header(alias="Idempotency-Key")], user=Depends(actor)):
        row = media.begin(user, body.file_name, body.mime_type, body.byte_size, idempotency_key, purpose="MOMENT_COVER")
        return {"id": row.id, "upload_url": f"/api/v1/moments/cover/uploads/{row.id}/content", "expires_at": row.expires_at}

    @router.put("/cover/uploads/{upload_id}/content", status_code=204)
    async def put_cover_upload_content(upload_id: str, request: Request, content_type: Annotated[str | None, Header(alias="Content-Type")] = None, user=Depends(actor)):
        content = bytearray()
        async for chunk in request.stream():
            content.extend(chunk)
        media.put_content(user, upload_id, bytes(content), (content_type or "").partition(";")[0].strip().casefold())
        return Response(status_code=204)

    @router.post("/cover/uploads/{upload_id}/complete")
    def complete_cover_upload(upload_id: str, idempotency_key: Annotated[str, Header(alias="Idempotency-Key")], user=Depends(actor)):
        row = media.complete(user, upload_id)
        media_url = avatar_storage.signed_read_url(row.object_key, 300) if row.status == "COMPLETED" and avatar_storage else None
        return {"id": row.id, "status": row.status, "media_url": media_url}

    @router.put("/cover")
    def set_cover(body: SetCover, idempotency_key: Annotated[str, Header(alias="Idempotency-Key")], user=Depends(actor)):
        return service.set_cover(user, body.upload_id, idempotency_key)

    @router.get("/{moment_id}")
    def detail(moment_id: str, user=Depends(actor)):
        return service.detail(user, moment_id)

    @router.delete("/{moment_id}", status_code=204)
    def delete(moment_id: str, idempotency_key: Annotated[str, Header(alias="Idempotency-Key")], user=Depends(actor)):
        service.delete(user, moment_id, idempotency_key)
        return Response(status_code=204)

    @router.post("/{moment_id}/likes", status_code=201)
    def like(moment_id: str, idempotency_key: Annotated[str, Header(alias="Idempotency-Key")], user=Depends(actor)):
        row = service.like(user, moment_id, idempotency_key)
        return {"id": row.id}

    @router.delete("/{moment_id}/likes", status_code=204)
    def unlike(moment_id: str, idempotency_key: Annotated[str, Header(alias="Idempotency-Key")], user=Depends(actor)):
        service.unlike(user, moment_id, idempotency_key)
        return Response(status_code=204)

    @router.post("/{moment_id}/comments", status_code=201)
    def comment(moment_id: str, body: Comment, idempotency_key: Annotated[str, Header(alias="Idempotency-Key")], user=Depends(actor)):
        row = service.comment(user, moment_id, body.text, body.parent_id, idempotency_key)

        with service.factory() as session:
            return service.comment_dto(session, row, user)

    @router.delete("/{moment_id}/comments/{comment_id}", status_code=204)
    def delete_comment(moment_id: str, comment_id: str, idempotency_key: Annotated[str, Header(alias="Idempotency-Key")], user=Depends(actor)):
        service.delete_comment(user, moment_id, comment_id, idempotency_key)
        return Response(status_code=204)

    @router.post("/{moment_id}/reports", status_code=201)
    def report(moment_id: str, body: Report, idempotency_key: Annotated[str, Header(alias="Idempotency-Key")], user=Depends(actor)):
        row = service.report(user, moment_id, body.reason_code, idempotency_key)
        return {"id": row.id, "status": row.status}

    return router